+++ create-vite-app/template-react/src/App.jsx
@@ -1,4 +1,4 @@
-import { React } from 'es-react'
+import React from 'react'
 import logo from './logo.svg'
 import './App.css'