+++ create-vite-app/template-react/src/main.jsx
@@ -1,4 +1,5 @@
-import { React, ReactDOM } from 'es-react'
+import React from 'react'
+import ReactDOM from 'react-dom'
 import './index.css'
 import App from './App'