+++ src/client/env.d.ts
@@ -1,0 +1,3 @@
+declare const __DEV__: boolean
+declare const __BASE__: string
+declare const __SW_ENABLED__: boolean