+++ src/node/server/serverPluginHmr.ts
@@ -29,6 +29,7 @@
 //    updates using the full paths of the dependencies.
 
 import { ServerPlugin } from '.'
+import fs from 'fs'
 import WebSocket from 'ws'
 import path from 'path'
 import chalk from 'chalk'
@@ -36,7 +37,6 @@ import hash_sum from 'hash-sum'
 import { SFCBlock } from '@vue/compiler-sfc'
 import { parseSFC, vueCache, srcImportMap } from './serverPluginVue'
 import { resolveImport } from './serverPluginModuleRewrite'
-import { cachedRead } from '../utils'
 import { FSWatcher } from 'chokidar'
 import MagicString from 'magic-string'
 import { parse } from '@babel/parser'
@@ -79,9 +79,11 @@ interface HMRPayload {
     | 'style-update'
     | 'style-remove'
     | 'full-reload'
+    | 'sw-bust-cache'
     | 'custom'
   timestamp: number
   path?: string
+  changeSrcPath?: string
   id?: string
   index?: number
   customData?: any
@@ -95,13 +97,16 @@ export const hmrPlugin: ServerPlugin = ({
   resolver,
   config
 }) => {
+  const hmrClient = fs.readFileSync(hmrClientFilePath, 'utf-8')
+
   app.use(async (ctx, next) => {
-    if (ctx.path !== hmrClientPublicPath) {
+    if (ctx.path === hmrClientPublicPath) {
+      ctx.type = 'js'
+      ctx.status = 200
+      ctx.body = hmrClient
+    } else {
       return next()
     }
-    debugHmr('serving hmr client')
-    ctx.type = 'js'
-    await cachedRead(ctx, hmrClientFilePath)
   })
 
   // start a websocket server to send hmr notifications to the client
@@ -179,6 +184,7 @@ export const hmrPlugin: ServerPlugin = ({
 
     // check which part of the file changed
     let needReload = false
+    let needCssModuleReload = false
     let needRerender = false
 
     if (!isEqual(descriptor.script, prevDescriptor.script)) {
@@ -194,16 +200,20 @@ export const hmrPlugin: ServerPlugin = ({
     const prevStyles = prevDescriptor.styles || []
     const nextStyles = descriptor.styles || []
     if (
-      (!needReload &&
-        prevStyles.some((s) => s.scoped) !==
-          nextStyles.some((s) => s.scoped)) ||
-      // TODO for now we force the component to reload on <style module> change
-      // but this should be optimizable to replace the __cssMoudles object
-      // on script and only trigger a rerender.
+      !needReload &&
+      prevStyles.some((s) => s.scoped) !== nextStyles.some((s) => s.scoped)
+    ) {
+      needReload = true
+    }
+
+    // css modules update causes a reload because the $style object is changed
+    // and it may be used in JS. It also needs to trigger a vue-style-update
+    // event so the client busts the sw cache.
+    if (
       prevStyles.some((s) => s.module != null) ||
       nextStyles.some((s) => s.module != null)
     ) {
-      needReload = true
+      needCssModuleReload = true
     }
 
     // only need to update styles if not reloading, since reload forces
@@ -234,7 +244,7 @@ export const hmrPlugin: ServerPlugin = ({
       })
     })
 
-    if (needReload) {
+    if (needReload || needCssModuleReload) {
       send({
         type: 'vue-reload',
         path: publicPath,
@@ -292,6 +302,7 @@ export const hmrPlugin: ServerPlugin = ({
       if (hasDeadEnd) {
         send({
           type: 'full-reload',
+          path: publicPath,
           timestamp
         })
         console.log(chalk.green(`[vite] `) + `page reloaded.`)
@@ -304,6 +315,7 @@ export const hmrPlugin: ServerPlugin = ({
           send({
             type: 'vue-reload',
             path: vueImporter,
+            changeSrcPath: publicPath,
             timestamp
           })
         })
@@ -315,6 +327,7 @@ export const hmrPlugin: ServerPlugin = ({
           send({
             type: 'js-update',
             path: jsImporter,
+            changeSrcPath: publicPath,
             timestamp
           })
         })
@@ -396,6 +409,7 @@ export function ensureMapEntry(map: HMRStateMap, key: string): Set<string> {
 }
 
 export function rewriteFileWithHMR(
+  root: string,
   source: string,
   importer: string,
   resolver: InternalResolver,
@@ -415,7 +429,7 @@ export function rewriteFileWithHMR(
 
   const registerDep = (e: StringLiteral) => {
     const deps = ensureMapEntry(hmrAcceptanceMap, importer)
-    const depPublicPath = resolveImport(importer, e.value, resolver)
+    const depPublicPath = resolveImport(root, importer, e.value, resolver)
     deps.add(depPublicPath)
     debugHmr(`        ${importer} accepts ${depPublicPath}`)
     ensureMapEntry(importerMap, depPublicPath).add(importer)