+++ src/node/server/serverPluginVue.ts
@@ -30,6 +30,7 @@ import { Context } from 'koa'
 import { transform } from '../esbuildService'
 import { InternalResolver } from '../resolver'
 import qs from 'querystring'
+import { seenUrls } from './serverPluginServeStatic'
 
 const debug = require('debug')('vite:sfc')
 const getEtag = require('etag')
@@ -47,11 +48,6 @@ export const vueCache = new LRUCache<string, CacheEntry>({
   max: 65535
 })
 
-const etagCacheCheck = (ctx: Context) => {
-  ctx.etag = getEtag(ctx.body)
-  ctx.status = ctx.etag === ctx.get('If-None-Match') ? 304 : 200
-}
-
 export const vuePlugin: ServerPlugin = ({
   root,
   app,
@@ -59,6 +55,18 @@ export const vuePlugin: ServerPlugin = ({
   watcher,
   config
 }) => {
+  const etagCacheCheck = (ctx: Context) => {
+    ctx.etag = getEtag(ctx.body)
+    // only add 304 tag check if not using service worker to cache user code
+    if (!config.serviceWorker) {
+      ctx.status =
+        seenUrls.has(ctx.url) && ctx.etag === ctx.get('If-None-Match')
+          ? 304
+          : 200
+      seenUrls.add(ctx.url)
+    }
+  }
+
   app.use(async (ctx, next) => {
     if (!ctx.path.endsWith('.vue') && !ctx.vue) {
       return next()